+++ rishi/__init__.py
@@ -1,4 +1,4 @@
-__version__ = "0.1.29"
+__version__ = "0.1.31"
 
 from urai import *
 from .core import *